refactor: extract runtime agent assembly
